+++ test/errorhandling.js
@@ -55,17 +55,17 @@ test('allow state changes in autorun', function(t) {
     var z = observable(3);
     
     m.autorun(function() {
-        if (x() !== 3)
-            z(x());
+        if (x.get() !== 3)
+            z.set(x.get());
     });
     
-    t.equal(x(), 3);
-    t.equal(z(), 3);
+    t.equal(x.get(), 3);
+    t.equal(z.get(), 3);
 
-    x(5); // autorunneres are allowed to change state
+    x.set(5); // autorunneres are allowed to change state
 
-    t.equal(x(), 5);
-    t.equal(z(), 5);
+    t.equal(x.get(), 5);
+    t.equal(z.get(), 5);
 
     t.equal(mobservable.extras.isComputingDerivation(), false);
     t.end();
@@ -98,14 +98,14 @@ test('allow array change in autorun', function(t) {
     var x = observable(3);
     var z = observable([]);
     var y = m.autorun(function() {
-        if (x() > 4)
-            z.push(x());
+        if (x.get() > 4)
+            z.push(x.get());
     });
     
-    x(5);
-    x(6);
+    x.set(5);
+    x.set(6);
     t.deepEqual(z.slice(), [5, 6])
-    x(2);
+    x.set(2);
     t.deepEqual(z.slice(), [5, 6])
 
     t.equal(mobservable.extras.isComputingDerivation(), false);
@@ -115,10 +115,10 @@ test('allow array change in autorun', function(t) {
 test('throw error if modification loop', function(t) {
     var x = observable(3);
     var dis = m.autorun(function() {
-        x(x() + 1); // is allowed to throw, but doesn't as the observables aren't bound yet during first execution
+        x.set(x.get() + 1); // is allowed to throw, but doesn't as the observables aren't bound yet during first execution
     });
     t.throws(() => {
-        x(5);
+        x.set(5);
     }, "Reaction doesn't converge to a stable state")
     t.end();
 })
@@ -132,32 +132,32 @@ test('cycle1', function(t) {
 })
 
 test('cycle2', function(t) {
-    var a = observable(function() { return b() * 2; });
-    var b = observable(function() { return a() * 2; });
+    var a = observable(function() { return b.get() * 2; });
+    var b = observable(function() { return a.get() * 2; });
     t.throws(() => {
-        b()
+        b.get()
     }, "Found cyclic dependency");
     t.end();
 })
 
 test('cycle3', function(t) {
-    var p = observable(function() { return p() * 2; });
+    var p = observable(function() { return p.get() * 2; });
     t.throws(() => {
-        p();
+        p.get();
     }, "Found cyclic dependency");
     t.end();
 })
 
 test('cycle3', function(t) {
     var z = observable(true);
-    var a = observable(function() { return z() ? 1 : b() * 2; });
-    var b = observable(function() { return a() * 2; });
+    var a = observable(function() { return z.get() ? 1 : b.get() * 2; });
+    var b = observable(function() { return a.get() * 2; });
 
     m.observe(b, voidObserver);
-    t.equal(1, a());
+    t.equal(1, a.get());
 
     t.throws(() => {
-        z(false); // introduces a cycle!
+        z.set(false); // introduces a cycle!
     }, "Found cyclic dependency");
     t.end();
 });
@@ -180,13 +180,13 @@ test('issue 86, converging cycles', function(t) {
     // should delete item 1 in first run, which works fine
     mobservable.autorun(() => {
         calcs++;
-        const i = findIndex(state.someArray, item => item.id === deleteThisId());
+        const i = findIndex(state.someArray, item => item.id === deleteThisId.get());
         state.someArray.remove(state.someArray[i]);
     });
     
     t.equal(state.someArray.length, 1); // should be 1, which prints fine
     t.equal(calcs, 1);
-    deleteThisId(2); // should delete item 2, but it errors on cycle
+    deleteThisId.set(2); // should delete item 2, but it errors on cycle
     
     t.equal(console.log(state.someArray.length, 0)); // should be 0, which never prints
     t.equal(calcs, 3);
@@ -198,10 +198,10 @@ test('slow converging cycle', function(t) {
     var x = mobservable.observable(1);
     var res = -1;
     mobservable.autorun(() => {
-        if (x() === 100)
-            res = x();
+        if (x.get() === 100)
+            res = x.get();
         else
-            x(x() + 1);
+            x.set(x.get() + 1);
     });
     
     // ideally the outcome should be 100 / 100.
@@ -209,11 +209,11 @@ test('slow converging cycle', function(t) {
     // is there a practical use case where such a pattern would be expected?
     // maybe we need to immediately register observers on the observable? but that would be slow....
     // or detect cycles and re-run the autorun in that case once?
-    t.equal(x(), 2)
+    t.equal(x.get(), 2)
     t.equal(res, -1);
     
-    x(7);
-    t.equal(x(), 100)
+    x.set(7);
+    t.equal(x.get(), 100)
     t.equal(res, 100);
     
     t.end();