+++ test/perf/perf.js
@@ -28,7 +28,7 @@ test('one observes ten thousand that observe one', function (t) {
     for (var i = 0; i < 10000; i++) {
         (function(idx) {
             observers.push(observable(function() {
-                return a() * idx;
+                return a.get() * idx;
             }))
         })(i);
     }
@@ -38,7 +38,7 @@ test('one observes ten thousand that observe one', function (t) {
     var b = observable(function() {
         var res = 0;
         for(var i = 0; i < observers.length; i++)
-        res += observers[i]();
+        	res += observers[i].get();
         bCalcs += 1;
         return res;
     })
@@ -46,11 +46,11 @@ test('one observes ten thousand that observe one', function (t) {
     var start = now();
 
     mobservable.observe(b, voidObserver, true); // start observers
-    t.equal(99990000, b());
+    t.equal(99990000, b.get());
     var initial = now();
 
-    a(3);
-    t.equal(149985000, b()); // yes, I verified ;-).
+    a.set(3);
+    t.equal(149985000, b.get()); // yes, I verified ;-).
     //t.equal(2, bCalcs);
     var end = now();
 
@@ -63,7 +63,7 @@ test('five hundrend properties that observe their sibling', function (t) {
     var observables = [observable(1)];
     for(var i = 0; i < 500; i++) {
         (function(idx) {
-            observables.push(observable(function() { return observables[idx]() + 1 }));
+            observables.push(observable(function() { return observables[idx].get() + 1 }));
         })(i);
     }
 
@@ -71,11 +71,11 @@ test('five hundrend properties that observe their sibling', function (t) {
 
     var last = observables[observables.length -1];
     mobservable.observe(last, voidObserver);
-    t.equal(501, last());
+    t.equal(501, last.get());
     var initial = now();
 
-    observables[0](2);
-    t.equal(502, last());
+    observables[0].set(2);
+    t.equal(502, last.get());
     var end = now();
 
     console.log("\n  Started/Updated in " + (initial - start) + "/" + (end - initial) + " ms.");
@@ -91,7 +91,7 @@ test('late dependency change', function(t) {
     var sum = observable(function() {
         var sum = 0;
         for(var i = 0; i < 100; i++)
-    	    sum += values[i]();
+    	    sum += values[i].get();
         return sum;
     })
 
@@ -100,9 +100,9 @@ test('late dependency change', function(t) {
     var start = new Date();
 
     for(var i = 0; i < 10000; i++)
-	    values[99](i);
+	    values[99].set(i);
 
-    t.equal(sum(), 9999);
+    t.equal(sum.get(), 9999);
     console.log("\n  Updated in " + ((new Date) - start) + "ms.");
     t.end();
 })
@@ -116,7 +116,7 @@ test('lots of unused computables', function(t) {
     for (var i = 0; i < 10000; i++) {
         (function(idx) {
             observers.push(observable(function() {
-                return a() * idx;
+                return a.get() * idx;
             }))
         })(i);
     }
@@ -125,7 +125,7 @@ test('lots of unused computables', function(t) {
     var b = observable(function() {
         var res = 0;
         for(var i = 0; i < observers.length; i++)
-        res += observers[i]();
+        	res += observers[i].get();
         return res;
     });
 
@@ -141,7 +141,7 @@ test('lots of unused computables', function(t) {
 
     var start = now();
 
-    a(3);
+    a.set(3);
     t.equal(sum, 49995000); // unchanged!
 
     var end = now();
@@ -155,13 +155,13 @@ test('many unreferenced observables', function(t) {
     var a = observable(3);
     var b = observable(6);
     var c = observable(7);
-    var d = observable(function() { return a() * b() * c() });
-    t.equal(d(), 126);
-    t.equal(d.$mobservable.isLazy, true);
+    var d = observable(function() { return a.get() * b.get() * c.get() });
+    t.equal(d.get(), 126);
+    t.equal(d.isLazy, true);
     var start = now();
     for(var i = 0; i < 10000; i++) {
-        c(i);
-        d();
+        c.set(i);
+        d.get();
     }
     var end = now();
 
@@ -179,7 +179,7 @@ test('array reduce', function(t) {
     var sum = observable(function() {
         aCalc++;
         return ar.reduce(function(a, c) {
-            return a + c * b();
+            return a + c * b.get();
         }, 0);
     });
     mobservable.observe(sum, voidObserver);
@@ -189,7 +189,7 @@ test('array reduce', function(t) {
     for(var i = 0; i < 1000; i++)
         ar.push(i);
 
-    t.equal(499500, sum());
+    t.equal(499500, sum.get());
     t.equal(1001, aCalc);
     aCalc = 0;
 
@@ -197,9 +197,9 @@ test('array reduce', function(t) {
 
     for(var i = 0; i < 1000; i++)
     ar[i] = ar[i] * 2;
-    b(2);
+    b.set(2);
 
-    t.equal(1998000, sum());
+    t.equal(1998000, sum.get());
     t.equal(1000, aCalc);
 
     var end = now();
@@ -217,7 +217,7 @@ test('array classic loop', function(t) {
         var s = 0;
         aCalc++;
         for(var i = 0; i < ar.length; i++)
-            s+=ar[i] * b();
+            s+=ar[i] * b.get();
         return s;
     });
     mobservable.observe(sum, voidObserver, true); // calculate
@@ -226,19 +226,19 @@ test('array classic loop', function(t) {
 
     t.equal(1, aCalc);
     for(var i = 0; i < 1000; i++)
-    ar.push(i);
+    	ar.push(i);
 
-    t.equal(499500, sum());
+    t.equal(499500, sum.get());
     t.equal(1001, aCalc);
 
     var initial = now();
     aCalc = 0;
 
     for(var i = 0; i < 1000; i++)
-    ar[i] = ar[i] * 2;
-    b(2);
+    	ar[i] = ar[i] * 2;
+    b.set(2);
 
-    t.equal(1998000, sum());
+    t.equal(1998000, sum.get());
     t.equal(1000, aCalc);
 
     var end = now();
@@ -256,15 +256,16 @@ function order_system_helper(t, usebatch, keepObserving) {
     var totalAmount = observable(function() {
         var sum = 0, l = orders.length;
         for(var i = 0; i < l; i++)
-        sum += orders[i].total();
+        	sum += orders[i].total.get();
         return sum;
     });
 
+	// TODO: use extendObservable!
     function OrderLine(order, price, amount) {
         this.price = observable(price);
         this.amount = observable(amount);
         this.total = observable(function() {
-            return order.vat() * this.price() * this.amount();
+            return order.vat.get() * this.price.get() * this.amount.get();
         }, this);
     }
 
@@ -273,14 +274,14 @@ function order_system_helper(t, usebatch, keepObserving) {
         this.lines = observable([]);
 
         this.vat = observable(function() {
-            if (this.includeVat())
-            return vat();
+            if (this.includeVat.get())
+            	return vat.get();
             return 1;
         }, this);
 
         this.total = observable(function() {
             return this.lines.reduce(function(acc, order) {
-                return acc + order.total();
+                return acc + order.total.get();
             }, 0);
         }, this);
     }
@@ -305,14 +306,14 @@ function order_system_helper(t, usebatch, keepObserving) {
     else
         setup();
 
-    t.equal(totalAmount(), 375000);
+    t.equal(totalAmount.get(), 375000);
 
     var initial = now();
 
     function update() {
         for(var i = 0; i < 50; i++)
-            orders[i].includeVat(!orders[i].includeVat());
-        vat(3);
+            orders[i].includeVat.set(!orders[i].includeVat.get());
+        vat.set(3);
     }
 
     if (usebatch)
@@ -320,7 +321,7 @@ function order_system_helper(t, usebatch, keepObserving) {
     else
         update();
 
-    t.equal(totalAmount(), 500000);
+    t.equal(totalAmount.get(), 500000);
 
     if (keepObserving)
         disp();