+++ test/typescript-tests.ts
@@ -32,9 +32,9 @@ class Order {
 
 test('observable', function(t) {
     var a = observable(3);
-    var b = observable(() => a() * 2);
+    var b = observable(() => a.get() * 2);
     t.ok(extras.getDNode(a));
-    t.equal(b(), 6);
+    t.equal(b.get(), 6);
     t.end();
 })